@@ -14,7 +14,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3",
+        "php": ">=7.0",
         "ext-curl": "*"
     },
     "require-dev": {
