@@ -9,11 +9,11 @@ use Curl\Curl;
 //     -d '{"id":"1","content":"Hello world!","date":"2015-06-30 19:42:21"}' \
 //     "https://httpbin.org/post"
 
-$data = array(
+$data = [
     'id' => '1',
     'content' => 'Hello world!',
     'date' => date('Y-m-d H:i:s'),
-);
+];
 
 $curl = new Curl();
 $curl->setDefaultJsonDecoder($assoc = true);
