@@ -16,7 +16,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @see next()
      * @see key()
      */
-    private $data = array();
+    private $data = [];
 
     /**
      * @var string[] Case-sensitive keys.
@@ -24,7 +24,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @see offsetUnset()
      * @see key()
      */
-    private $keys = array();
+    private $keys = [];
 
     /**
      * Construct
