@@ -9,8 +9,13 @@ class Test
     const TEST_URL = 'http://127.0.0.1:8000/';
     const ERROR_URL = 'http://1.2.3.4/';
 
-    public function __construct()
+    public $message;
+
+    private $testUrl;
+
+    public function __construct($port = null)
     {
+        $this->testUrl = $port === null ? self::TEST_URL : $this->getTestUrl($port);
         $this->curl = new Curl();
         $this->curl->setOpt(CURLOPT_SSL_VERIFYPEER, false);
         $this->curl->setOpt(CURLOPT_SSL_VERIFYHOST, false);
@@ -21,12 +26,22 @@ class Test
         $this->curl->setHeader('X-DEBUG-TEST', $test);
         $request_method = strtolower($request_method);
         if ($arg1 !== null && $arg2 !== null) {
-            $this->curl->$request_method(self::TEST_URL, $arg1, $arg2);
+            $this->curl->$request_method($this->testUrl, $arg1, $arg2);
         } elseif ($arg1 !== null) {
-            $this->curl->$request_method(self::TEST_URL, $arg1);
+            $this->curl->$request_method($this->testUrl, $arg1);
         } else {
-            $this->curl->$request_method(self::TEST_URL);
+            $this->curl->$request_method($this->testUrl);
         }
+
+        $this->message =
+            'error message: ' . $this->curl->errorMessage . "\n" .
+            'curl error message: ' . $this->curl->curlErrorMessage . "\n" .
+            'http error message: ' . $this->curl->httpErrorMessage . "\n" .
+            'error code: ' . $this->curl->errorCode . "\n" .
+            'curl error code: ' . $this->curl->curlErrorCode . "\n" .
+            'raw response: ' . $this->curl->rawResponse . "\n" .
+            '';
+
         return $this->curl->response;
     }
 
@@ -45,7 +60,7 @@ class Test
         \PHPUnit\Framework\Assert::assertEquals($request_method, $this->curl->responseHeaders['X-REQUEST-METHOD']);
     }
 
-    public function chainRequests($first, $second, $data = array())
+    public function chainRequests($first, $second, $data = [])
     {
         $this->chainedRequest($first, $data);
         $this->chainedRequest($second, $data);
@@ -53,12 +68,15 @@ class Test
 
     public static function getTestUrl($port)
     {
-        if (getenv('PHP_CURL_CLASS_LOCAL_TEST') === 'yes' ||
-            in_array(getenv('TRAVIS_PHP_VERSION'), array('7.0', '7.1', '7.2', '7.3', '7.4', 'nightly'))) {
-            return 'http://127.0.0.1:' . $port . '/';
-        } else {
-            return self::TEST_URL;
-        }
+        // Return url pointing to a test server running on the specified port.
+        // To avoid installing and configuring a web server for the tests, PHP's
+        // built-in development server is used. As each development server can
+        // only handle one request at a time (single-threaded) and some tests
+        // expect the server to handle requests simultaneously, multiple
+        // instances are run on different ports. With this setup, requests in
+        // the test can be made to the various port urls without having to be
+        // handled sequentially.
+        return 'http://127.0.0.1:' . $port . '/';
     }
 }
 
@@ -114,9 +132,9 @@ if (function_exists('finfo_open')) {
 
 function upload_file_to_server($upload_file_path) {
     $upload_test = new Test();
-    $upload_test->server('upload_response', 'POST', array(
+    $upload_test->server('upload_response', 'POST', [
         'image' => '@' . $upload_file_path,
-    ));
+    ]);
     $uploaded_file_path = $upload_test->curl->response->file_path;
 
     // Ensure files are not the same path.
@@ -132,9 +150,9 @@ function remove_file_from_server($uploaded_file_path) {
     $download_test = new Test();
 
     // Ensure file successfully removed.
-    assert('true' === $download_test->server('upload_cleanup', 'POST', array(
+    assert('true' === $download_test->server('upload_cleanup', 'POST', [
         'file_path' => $uploaded_file_path,
-    )));
+    ]));
     assert(file_exists($uploaded_file_path) === false);
 }
 
@@ -152,4 +170,28 @@ function get_multi_curl_property_value($instance, $property_name)
     $property = $reflector->getProperty($property_name);
     $property->setAccessible(true);
     return $property->getValue($instance);
+}
+
+function get_request_stats($request_stats, $multi_curl)
+{
+    $messages =
+        ['total duration: ' . sprintf('%.6f', round($multi_curl->stopTime - $multi_curl->startTime, 6)) . "\n"];
+
+    foreach ($request_stats as $instance_id => &$value) {
+        $value['relative_start'] = sprintf('%.6f', round($value['start'] - $multi_curl->startTime, 6));
+        $value['relative_stop'] = sprintf('%.6f', round($value['stop'] - $multi_curl->startTime, 6));
+        $value['duration'] = (string)round($value['stop'] - $value['start'], 6);
+
+        $messages[] =
+            $value['relative_start']        . ' - ' . 'request ' . $instance_id . ' start'        . "\n" .
+            $value['relative_stop']         . ' - ' . 'request ' . $instance_id . ' complete'     . "\n" .
+            $value['duration']              . ' - ' . 'request ' . $instance_id . ' duration'     . "\n";
+
+        unset($value['start']);
+        unset($value['stop']);
+    }
+
+    $request_stats['message'] = implode("\n", $messages);
+
+    return $request_stats;
 }
