@@ -669,8 +669,8 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
             \PHPUnit\Framework\Assert::assertTrue($instance->error);
             \PHPUnit\Framework\Assert::assertTrue($instance->curlError);
             \PHPUnit\Framework\Assert::assertFalse($instance->httpError);
-            $possible_errors = array(
-                CURLE_SEND_ERROR, CURLE_OPERATION_TIMEOUTED, CURLE_COULDNT_CONNECT, CURLE_GOT_NOTHING);
+            $possible_errors = [
+                CURLE_SEND_ERROR, CURLE_OPERATION_TIMEOUTED, CURLE_COULDNT_CONNECT, CURLE_GOT_NOTHING];
             \PHPUnit\Framework\Assert::assertTrue(
                 in_array($instance->errorCode, $possible_errors, true),
                 'errorCode: ' . $instance->errorCode
@@ -2045,7 +2045,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
     {
         $multi_curl_user_agent = 'multi curl user agent';
         $curl_user_agent = 'curl user agent';
-        $data = array('key' => 'HTTP_USER_AGENT');
+        $data = ['key' => 'HTTP_USER_AGENT'];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'server');
@@ -2250,15 +2250,15 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
     {
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'setcookie');
-        $multi_curl->setCookies(array(
+        $multi_curl->setCookies([
             'mycookie' => 'yum',
             'cookie-for-all-before' => 'a',
-        ));
+        ]);
 
         $get_1 = $multi_curl->addGet(Test::TEST_URL);
-        $get_1->setCookies(array(
+        $get_1->setCookies([
             'cookie-for-1st-request' => '1',
-        ));
+        ]);
         $get_1->complete(function ($instance) {
             \PHPUnit\Framework\Assert::assertEquals('yum', $instance->responseCookies['mycookie']);
             \PHPUnit\Framework\Assert::assertEquals('a', $instance->responseCookies['cookie-for-all-before']);
@@ -2267,13 +2267,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         });
 
         $get_2 = $multi_curl->addGet(Test::TEST_URL);
-        $get_2->setCookies(array(
+        $get_2->setCookies([
             'cookie-for-2nd-request' => '2',
-        ));
+        ]);
         $get_2->beforeSend(function ($instance) {
-            $instance->setCookies(array(
+            $instance->setCookies([
                 'mycookie' => 'yummy',
-            ));
+            ]);
         });
         $get_2->complete(function ($instance) {
             \PHPUnit\Framework\Assert::assertEquals('yummy', $instance->responseCookies['mycookie']);
@@ -2282,9 +2282,9 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
             \PHPUnit\Framework\Assert::assertEquals('2', $instance->responseCookies['cookie-for-2nd-request']);
         });
 
-        $multi_curl->setCookies(array(
+        $multi_curl->setCookies([
             'cookie-for-all-after' => 'b',
-        ));
+        ]);
         $multi_curl->start();
 
         $this->assertEquals('yum', $get_1->responseCookies['mycookie']);
@@ -2293,25 +2293,25 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
     public function testJsonRequest()
     {
-        foreach (array(
-                array(
-                    array(
+        foreach ([
+                [
+                    [
                         'key' => 'value',
-                    ),
+                    ],
                     '{"key":"value"}',
-                ),
-                array(
-                    array(
+                ],
+                [
+                    [
                         'key' => 'value',
-                        'strings' => array(
+                        'strings' => [
                             'a',
                             'b',
                             'c',
-                        ),
-                    ),
+                        ],
+                    ],
                     '{"key":"value","strings":["a","b","c"]}',
-                ),
-            ) as $test) {
+                ],
+            ] as $test) {
             list($data, $expected_response) = $test;
 
             $multi_curl = new MultiCurl();
@@ -2322,18 +2322,18 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
             $multi_curl->addPost(Test::TEST_URL, json_encode($data));
             $multi_curl->start();
 
-            foreach (array(
+            foreach ([
                 'Content-Type',
                 'content-type',
-                'CONTENT-TYPE') as $key) {
-                foreach (array(
+                'CONTENT-TYPE'] as $key) {
+                foreach ([
                     'APPLICATION/JSON',
                     'APPLICATION/JSON; CHARSET=UTF-8',
                     'APPLICATION/JSON;CHARSET=UTF-8',
                     'application/json',
                     'application/json; charset=utf-8',
                     'application/json;charset=UTF-8',
-                    ) as $value) {
+                    ] as $value) {
                     $multi_curl = new MultiCurl();
                     $multi_curl->setHeader('X-DEBUG-TEST', 'post_json');
                     $multi_curl->setHeader($key, $value);
@@ -2358,10 +2358,10 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
     public function testJsonDecoder()
     {
-        $data = array(
+        $data = [
             'key' => 'Content-Type',
             'value' => 'application/json',
-        );
+        ];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'json_response');
@@ -2571,9 +2571,9 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $downloaded_file_path = tempnam('/tmp', 'php-curl-class.');
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'download_response');
-        $multi_curl->addDownload(Test::TEST_URL . '?' . http_build_query(array(
+        $multi_curl->addDownload(Test::TEST_URL . '?' . http_build_query([
             'file_path' => $uploaded_file_path,
-        )), $downloaded_file_path);
+        ]), $downloaded_file_path);
         $multi_curl->complete(function ($instance) use ($upload_file_path) {
             \PHPUnit\Framework\Assert::assertFalse($instance->error);
             \PHPUnit\Framework\Assert::assertEquals(md5_file($upload_file_path), $instance->responseHeaders['ETag']);
@@ -2601,7 +2601,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
         $filesize = filesize($filename);
 
-        foreach (array(
+        foreach ([
                 false,
                 0,
                 1,
@@ -2632,7 +2632,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
                 $filesize + 2,
                 $filesize + 3,
 
-            ) as $length) {
+            ] as $length) {
             $source = Test::TEST_URL;
             $destination = \Helper\get_tmp_file_path();
 
@@ -2669,9 +2669,9 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
             // Download (the remaining bytes of) the file.
             $multi_curl = new MultiCurl();
             $multi_curl->setHeader('X-DEBUG-TEST', 'download_file_range');
-            $multi_curl->addDownload($source . '?' . http_build_query(array(
+            $multi_curl->addDownload($source . '?' . http_build_query([
                 'file_path' => $uploaded_file_path,
-            )), $destination);
+            ]), $destination);
 
             clearstatcache();
 
@@ -2726,18 +2726,18 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         // Upload a file.
         $upload_file_path = \Helper\get_png();
         $upload_test = new Test();
-        $upload_test->server('upload_response', 'POST', array(
+        $upload_test->server('upload_response', 'POST', [
             'image' => '@' . $upload_file_path,
-        ));
+        ]);
         $uploaded_file_path = $upload_test->curl->response->file_path;
 
         // Download the file.
         $download_callback_called = false;
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'download_response');
-        $multi_curl->addDownload(Test::TEST_URL . '?' . http_build_query(array(
+        $multi_curl->addDownload(Test::TEST_URL . '?' . http_build_query([
             'file_path' => $uploaded_file_path,
-        )), function ($instance, $fh) use (&$download_callback_called) {
+        ]), function ($instance, $fh) use (&$download_callback_called) {
             \PHPUnit\Framework\Assert::assertFalse($download_callback_called);
             \PHPUnit\Framework\Assert::assertInstanceOf('Curl\Curl', $instance);
             \PHPUnit\Framework\Assert::assertTrue(is_resource($fh));
@@ -2751,9 +2751,9 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($download_callback_called);
 
         // Remove server file.
-        $this->assertEquals('true', $upload_test->server('upload_cleanup', 'POST', array(
+        $this->assertEquals('true', $upload_test->server('upload_cleanup', 'POST', [
             'file_path' => $uploaded_file_path,
-        )));
+        ]));
 
         unlink($upload_file_path);
         $this->assertFalse(file_exists($upload_file_path));
@@ -2779,11 +2779,11 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
     public function testSetUrlInConstructor()
     {
-        $data = array('key' => 'value');
+        $data = ['key' => 'value'];
 
         $multi_curl = new MultiCurl(Test::TEST_URL);
         $multi_curl->setHeader('X-DEBUG-TEST', 'delete_with_body');
-        $multi_curl->addDelete($data, array('wibble' => 'wubble'))->complete(function ($instance) {
+        $multi_curl->addDelete($data, ['wibble' => 'wubble'])->complete(function ($instance) {
             \PHPUnit\Framework\Assert::assertEquals(
                 '{"get":{"key":"value"},"delete":{"wibble":"wubble"}}',
                 $instance->rawResponse
@@ -2853,7 +2853,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
     public function testSetUrl()
     {
-        $data = array('key' => 'value');
+        $data = ['key' => 'value'];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setUrl(Test::TEST_URL);
@@ -2943,15 +2943,15 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
     {
         $multi_curl = new MultiCurl();
 
-        $urls = array();
-        $copy_of_urls = array();
+        $urls = [];
+        $copy_of_urls = [];
         for ($i = 0; $i < 10; $i++) {
             $url = Test::TEST_URL . '?' . md5(mt_rand());
             $urls[] = $url;
             $copy_of_urls[] = $url;
         }
 
-        $urls_called = array();
+        $urls_called = [];
         $multi_curl->complete(function ($instance) use (&$multi_curl, &$urls, &$urls_called) {
             $urls_called[] = $instance->url;
             $next_url = array_pop($urls);
@@ -2977,9 +2977,22 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl = new MultiCurl();
         $multi_curl->addGet(Test::TEST_URL);
         $multi_curl->start();
-        $this->assertTrue(is_resource($multi_curl->multiCurl));
+        $this->assertNotNull($multi_curl->multiCurl);
         $multi_curl->close();
-        $this->assertFalse(is_resource($multi_curl->multiCurl));
+        $this->assertNull($multi_curl->multiCurl);
+    }
+
+    public function testCookieJarAfterClose()
+    {
+        $cookie_jar = tempnam('/tmp', 'php-curl-class.');
+
+        $multi_curl = new MultiCurl();
+        $multi_curl->setCookieJar($cookie_jar);
+        $multi_curl->addGet(Test::TEST_URL);
+        $multi_curl->start();
+        $multi_curl->close();
+        $cookies = file_get_contents($cookie_jar);
+        $this->assertNotEmpty($cookies);
     }
 
     public function testMultiPostRedirectGet()
@@ -2988,7 +3001,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl = new MultiCurl(Test::TEST_URL);
         $multi_curl->setOpt(CURLOPT_FOLLOWLOCATION, true);
         $multi_curl->setHeader('X-DEBUG-TEST', 'post_redirect_get');
-        $multi_curl->addPost(array(), false)->complete(function ($instance) {
+        $multi_curl->addPost([], false)->complete(function ($instance) {
             \PHPUnit\Framework\Assert::assertEquals('Redirected: POST', $instance->response);
         });
         $multi_curl->start();
@@ -2997,7 +3010,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl = new MultiCurl(Test::TEST_URL);
         $multi_curl->setOpt(CURLOPT_FOLLOWLOCATION, true);
         $multi_curl->setHeader('X-DEBUG-TEST', 'post_redirect_get');
-        $multi_curl->addPost(array(), true)->complete(function ($instance) {
+        $multi_curl->addPost([], true)->complete(function ($instance) {
             \PHPUnit\Framework\Assert::assertEquals('Redirected: GET', $instance->response);
         });
         $multi_curl->start();
@@ -3024,10 +3037,10 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
     {
         $request_key = 'X-Request-Id';
         $request_value = '1';
-        $data = array(
+        $data = [
             'test' => 'server',
             'key' => 'HTTP_X_REQUEST_ID',
-        );
+        ];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader($request_key, $request_value);
@@ -3094,59 +3107,59 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->start();
     }
 
-    public function testRetryMulti()
+    public function testRetry()
     {
-        $tests = array(
-            array(
+        $tests = [
+            [
                 'maximum_number_of_retries' => null,
                 'failures' => 0,
                 'expect_success' => true,
                 'expect_attempts' => 1,
                 'expect_retries' => 0,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 0,
                 'failures' => 0,
                 'expect_success' => true,
                 'expect_attempts' => 1,
                 'expect_retries' => 0,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 0,
                 'failures' => 1,
                 'expect_success' => false,
                 'expect_attempts' => 1,
                 'expect_retries' => 0,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 1,
                 'failures' => 1,
                 'expect_success' => true,
                 'expect_attempts' => 2,
                 'expect_retries' => 1,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 1,
                 'failures' => 2,
                 'expect_success' => false,
                 'expect_attempts' => 2,
                 'expect_retries' => 1,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 2,
                 'failures' => 2,
                 'expect_success' => true,
                 'expect_attempts' => 3,
                 'expect_retries' => 2,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 3,
                 'failures' => 3,
                 'expect_success' => true,
                 'expect_attempts' => 4,
                 'expect_retries' => 3,
-            ),
-        );
+            ],
+        ];
         foreach ($tests as $test) {
             $maximum_number_of_retries = $test['maximum_number_of_retries'];
             $failures = $test['failures'];
@@ -3162,7 +3175,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
                 $multi_curl->setRetry($maximum_number_of_retries);
             }
 
-            $instance = $multi_curl->addGet(Test::TEST_URL, array('failures' => $failures));
+            $instance = $multi_curl->addGet(Test::TEST_URL, ['failures' => $failures]);
             $multi_curl->start();
 
             $this->assertEquals($expect_success, !$instance->error);
@@ -3171,59 +3184,59 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         }
     }
 
-    public function testRetryCallableMulti()
+    public function testRetryCallable()
     {
-        $tests = array(
-            array(
+        $tests = [
+            [
                 'maximum_number_of_retries' => null,
                 'failures' => 0,
                 'expect_success' => true,
                 'expect_attempts' => 1,
                 'expect_retries' => 0,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 0,
                 'failures' => 0,
                 'expect_success' => true,
                 'expect_attempts' => 1,
                 'expect_retries' => 0,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 0,
                 'failures' => 1,
                 'expect_success' => false,
                 'expect_attempts' => 1,
                 'expect_retries' => 0,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 1,
                 'failures' => 1,
                 'expect_success' => true,
                 'expect_attempts' => 2,
                 'expect_retries' => 1,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 1,
                 'failures' => 2,
                 'expect_success' => false,
                 'expect_attempts' => 2,
                 'expect_retries' => 1,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 2,
                 'failures' => 2,
                 'expect_success' => true,
                 'expect_attempts' => 3,
                 'expect_retries' => 2,
-            ),
-            array(
+            ],
+            [
                 'maximum_number_of_retries' => 3,
                 'failures' => 3,
                 'expect_success' => true,
                 'expect_attempts' => 4,
                 'expect_retries' => 3,
-            ),
-        );
+            ],
+        ];
         foreach ($tests as $test) {
             $maximum_number_of_retries = $test['maximum_number_of_retries'];
             $failures = $test['failures'];
@@ -3242,12 +3255,129 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
                 });
             }
 
-            $instance = $multi_curl->addGet(Test::TEST_URL, array('failures' => $failures));
+            $instance = $multi_curl->addGet(Test::TEST_URL, ['failures' => $failures]);
             $multi_curl->start();
 
             $this->assertEquals($expect_success, !$instance->error);
             $this->assertEquals($expect_attempts, $instance->attempts);
             $this->assertEquals($expect_retries, $instance->retries);
+        }
+    }
+
+    public function testRelativeUrl()
+    {
+        $multi_curl = new MultiCurl(Test::TEST_URL . 'path/');
+        $this->assertEquals('http://127.0.0.1:8000/path/', (string)$multi_curl->baseUrl);
+
+        $get_1 = $multi_curl->addGet('test', [
+            'a' => '1',
+            'b' => '2',
+        ]);
+        $this->assertEquals('http://127.0.0.1:8000/path/test?a=1&b=2', (string)$get_1->url);
+
+        $get_2 = $multi_curl->addGet('/root', [
+            'c' => '3',
+            'd' => '4',
+        ]);
+        $this->assertEquals('http://127.0.0.1:8000/root?c=3&d=4', (string)$get_2->url);
+
+        $tests = [
+            [
+                'args' => [
+                    'http://www.example.com/',
+                    '/foo',
+                ],
+                'expected' => 'http://www.example.com/foo',
+            ],
+            [
+                'args' => [
+                    'http://www.example.com/',
+                    '/foo/',
+                ],
+                'expected' => 'http://www.example.com/foo/',
+            ],
+            [
+                'args' => [
+                    'http://www.example.com/',
+                    '/dir/page.html',
+                ],
+                'expected' => 'http://www.example.com/dir/page.html',
+            ],
+            [
+                'args' => [
+                    'http://www.example.com/dir1/page2.html',
+                    '/dir/page.html',
+                ],
+                'expected' => 'http://www.example.com/dir/page.html',
+            ],
+            [
+                'args' => [
+                    'http://www.example.com/dir1/page2.html',
+                    'dir/page.html',
+                ],
+                'expected' => 'http://www.example.com/dir1/dir/page.html',
+            ],
+            [
+                'args' => [
+                    'http://www.example.com/dir1/dir3/page.html',
+                    '../dir/page.html',
+                ],
+                'expected' => 'http://www.example.com/dir1/dir/page.html',
+            ],
+        ];
+        foreach ($tests as $test) {
+            $multi_curl = new MultiCurl($test['args']['0']);
+            $multi_curl->setUrl($test['args']['1']);
+            $this->assertEquals(
+                $test['expected'],
+                $multi_curl->getOpt(CURLOPT_URL),
+                "Joint URLs: '{$test['args']['0']}', '{$test['args']['1']}'"
+            );
+
+            $multi_curl = new MultiCurl($test['args']['0']);
+            $multi_curl->setUrl($test['args']['1'], ['a' => '1', 'b' => '2']);
+            $this->assertEquals(
+                $test['expected'] . '?a=1&b=2',
+                $multi_curl->getOpt(CURLOPT_URL),
+                "Joint URL '{$test['args']['0']}' with parameters a=1, b=2"
+            );
+
+            $multi_curl = new MultiCurl();
+            $multi_curl->setUrl($test['args']['0']);
+            $multi_curl->setUrl($test['args']['1']);
+            $this->assertEquals(
+                $test['expected'],
+                $multi_curl->getOpt(CURLOPT_URL),
+                "Joint URLs: '{$test['args']['0']}', '{$test['args']['1']}'"
+            );
+
+            $multi_curl = new MultiCurl();
+            $multi_curl->setUrl($test['args']['0'], ['a' => '1', 'b' => '2']);
+            $multi_curl->setUrl($test['args']['1']);
+            $this->assertEquals(
+                $test['expected'],
+                $multi_curl->getOpt(CURLOPT_URL),
+                "Joint URL '{$test['args']['0']}' with parameters a=1, b=2 and URL '{$test['args']['1']}'"
+            );
+
+            $multi_curl = new MultiCurl();
+            $multi_curl->setUrl($test['args']['0']);
+            $multi_curl->setUrl($test['args']['1'], ['a' => '1', 'b' => '2']);
+            $this->assertEquals(
+                $test['expected'] . '?a=1&b=2',
+                $multi_curl->getOpt(CURLOPT_URL),
+                "Joint URL '{$test['args']['0']}' and URL '{$test['args']['1']}' with parameters a=1, b=2"
+            );
+
+            $multi_curl = new MultiCurl();
+            $multi_curl->setUrl($test['args']['0'], ['a' => '1', 'b' => '2']);
+            $multi_curl->setUrl($test['args']['1'], ['c' => '3', 'd' => '4']);
+            $this->assertEquals(
+                $test['expected'] . '?c=3&d=4',
+                $multi_curl->getOpt(CURLOPT_URL),
+                "Joint URL '{$test['args']['0']}' with parameters a=1, b=2 " .
+                "and URL '{$test['args']['1']}' with parameters c=3, d=4"
+            );
         }
     }
 
@@ -3312,13 +3442,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
     public function testSetProxiesRandomProxy()
     {
-        $proxies = array(
+        $proxies = [
             'example.com:80',
             'example.com:443',
             'example.com:1080',
             'example.com:3128',
             'example.com:8080',
-        );
+        ];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setProxies($proxies);
@@ -3351,13 +3481,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
     public function testSetProxiesAlreadySet()
     {
-        $proxies = array(
+        $proxies = [
             'example.com:80',
             'example.com:443',
             'example.com:1080',
             'example.com:3128',
             'example.com:8080',
-        );
+        ];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setProxies($proxies);
@@ -3424,138 +3554,138 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
 
     public function testSetRateLimitUnits()
     {
-        foreach (array(
-                array(
+        foreach ([
+                [
                     'rate_limit' => '1/s',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '1/1s',
                         'max_requests' => '1',
                         'interval' => '1',
                         'unit' => 's',
                         'interval_seconds' => '1',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '1/1s',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '1/1s',
                         'max_requests' => '1',
                         'interval' => '1',
                         'unit' => 's',
                         'interval_seconds' => '1',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '10/60s',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '10/60s',
                         'max_requests' => '10',
                         'interval' => '60',
                         'unit' => 's',
                         'interval_seconds' => '60',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '10/60S',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '10/60s',
                         'max_requests' => '10',
                         'interval' => '60',
                         'unit' => 's',
                         'interval_seconds' => '60',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '1/m',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '1/1m',
                         'max_requests' => '1',
                         'interval' => '1',
                         'unit' => 'm',
                         'interval_seconds' => '60',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '1/1m',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '1/1m',
                         'max_requests' => '1',
                         'interval' => '1',
                         'unit' => 'm',
                         'interval_seconds' => '60',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '5000/60m',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '5000/60m',
                         'max_requests' => '5000',
                         'interval' => '60',
                         'unit' => 'm',
                         'interval_seconds' => '3600',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '5000/60M',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '5000/60m',
                         'max_requests' => '5000',
                         'interval' => '60',
                         'unit' => 'm',
                         'interval_seconds' => '3600',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '1/h',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '1/1h',
                         'max_requests' => '1',
                         'interval' => '1',
                         'unit' => 'h',
                         'interval_seconds' => '3600',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '1/1h',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '1/1h',
                         'max_requests' => '1',
                         'interval' => '1',
                         'unit' => 'h',
                         'interval_seconds' => '3600',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '5000/1h',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '5000/1h',
                         'max_requests' => '5000',
                         'interval' => '1',
                         'unit' => 'h',
                         'interval_seconds' => '3600',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '100000/24h',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '100000/24h',
                         'max_requests' => '100000',
                         'interval' => '24',
                         'unit' => 'h',
                         'interval_seconds' => '86400',
-                    ),
-                ),
-                array(
+                    ],
+                ],
+                [
                     'rate_limit' => '100000/24H',
-                    'expected' => array(
+                    'expected' => [
                         'rate_limit' => '100000/24h',
                         'max_requests' => '100000',
                         'interval' => '24',
                         'unit' => 'h',
                         'interval_seconds' => '86400',
-                    ),
-                ),
-            ) as $test) {
+                    ],
+                ],
+            ] as $test) {
             $multi_curl = new MultiCurl();
             $multi_curl->setRateLimit($test['rate_limit']);
 
@@ -3593,16 +3723,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R4--|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -3615,51 +3742,43 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=1');
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=1');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 1.
-        $this->assertGreaterThanOrEqual(0.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(1.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(0.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(1.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 1.
-        $this->assertGreaterThanOrEqual(0.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(1.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(0.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(1.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 6.
-        $this->assertGreaterThanOrEqual(5.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(6.5, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(5.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(6.5, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 6.
-        $this->assertGreaterThanOrEqual(5.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(6.5, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(5.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(6.5, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(11.5, $request_stats['4']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecond2()
@@ -3673,16 +3792,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R4--|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -3695,51 +3811,43 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=2');
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=1');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 1.
-        $this->assertGreaterThanOrEqual(0.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(1.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(0.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(1.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 2.
-        $this->assertGreaterThanOrEqual(1.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(2.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(1.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(2.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 6.
-        $this->assertGreaterThanOrEqual(5.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(6.5, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(5.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(6.5, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(11.5, $request_stats['4']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecond3()
@@ -3751,16 +3859,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R4--|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -3773,51 +3878,43 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=5');
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=1');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 2.
-        $this->assertGreaterThanOrEqual(1.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(2.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(1.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(2.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 5.
-        $this->assertGreaterThanOrEqual(4.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(4.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 10.
-        $this->assertGreaterThanOrEqual(9.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(9.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(11.5, $request_stats['4']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecond4()
@@ -3829,16 +3926,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R4--|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -3851,51 +3945,43 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=6');
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=1');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start'], $request_stats['message']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start'], $request_stats['message']);
         // Assert R0 ends around 2.
-        $this->assertGreaterThanOrEqual(1.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(2.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(1.8, $request_stats['0']['relative_stop'], $request_stats['message']);
+        $this->assertLessThanOrEqual(2.5, $request_stats['0']['relative_stop'], $request_stats['message']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start'], $request_stats['message']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start'], $request_stats['message']);
         // Assert R1 ends around 9.
-        $this->assertGreaterThanOrEqual(8.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(9.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(8.8, $request_stats['1']['relative_stop'], $request_stats['message']);
+        $this->assertLessThanOrEqual(9.5, $request_stats['1']['relative_stop'], $request_stats['message']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start'], $request_stats['message']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start'], $request_stats['message']);
         // Assert R2 ends around 8.
-        $this->assertGreaterThanOrEqual(7.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(8.5 + 1, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(7.8, $request_stats['2']['relative_stop'], $request_stats['message']);
+        $this->assertLessThanOrEqual(8.5 + 1, $request_stats['2']['relative_stop'], $request_stats['message']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start'], $request_stats['message']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start'], $request_stats['message']);
         // Assert R3 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['3']['relative_stop'], $request_stats['message']);
+        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['3']['relative_stop'], $request_stats['message']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start'], $request_stats['message']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start'], $request_stats['message']);
         // Assert R4 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['relative_stop'], $request_stats['message']);
+        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['4']['relative_stop'], $request_stats['message']);
     }
 
     public function testSetRateLimitPerSecond5()
@@ -3908,16 +3994,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R4------|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -3930,51 +4013,43 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=2');
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=2');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5 + 1, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5 + 1, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5 + 1, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5 + 1, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5, $request_stats['4']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecond6()
@@ -3987,16 +4062,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R4------|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4009,51 +4081,43 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=2');
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=2');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 9.
-        $this->assertGreaterThanOrEqual(8.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(9.5 + 1, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(8.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(9.5 + 1, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 7.
-        $this->assertGreaterThanOrEqual(6.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(7.5 + 1, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(6.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(7.5 + 1, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5, $request_stats['4']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecond7()
@@ -4066,16 +4130,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R5------|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4089,58 +4150,50 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=2');
         $multi_curl->addGet(Test::getTestUrl('8006') . '?seconds=2');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 2.
-        $this->assertGreaterThanOrEqual(1.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(2.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(1.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(2.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(11.5 + 1, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['4']['relative_stop']);
 
         // Assert R5 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['5']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['5']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['5']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['5']['relative_start']);
         // Assert R5 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['5']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['5']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['5']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['5']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecond8()
@@ -4153,16 +4206,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R5------|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4176,58 +4226,50 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=1');
         $multi_curl->addGet(Test::getTestUrl('8006') . '?seconds=2');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 8.
-        $this->assertGreaterThanOrEqual(7.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(8.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(7.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(8.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 9.
-        $this->assertGreaterThanOrEqual(8.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(9.5 + 1, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(8.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(9.5 + 1, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 11.
-        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(11.5, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(10.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(11.5, $request_stats['4']['relative_stop']);
 
         // Assert R5 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['5']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['5']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['5']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['5']['relative_start']);
         // Assert R5 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['5']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5, $request_stats['5']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['5']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5, $request_stats['5']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecond9()
@@ -4240,16 +4282,13 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         //                                          R5------|
         //  0   1   2   3   4   5   6   7   8   9   10  11  12
 
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('2/5s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4263,71 +4302,60 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8005') . '?seconds=2');
         $multi_curl->addGet(Test::getTestUrl('8006') . '?seconds=2');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R0 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['0']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5, $request_stats['0']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['0']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5, $request_stats['0']['relative_stop']);
 
         // Assert R1 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['1']['relative_start']);
         // Assert R1 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['1']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5, $request_stats['1']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['1']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5, $request_stats['1']['relative_stop']);
 
         // Assert R2 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['2']['relative_start']);
         // Assert R2 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['2']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['2']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['2']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['2']['relative_stop']);
 
         // Assert R3 starts around 5 and not before.
-        $this->assertGreaterThanOrEqual(5, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(5.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(5, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(5.5, $request_stats['3']['relative_start']);
         // Assert R3 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['3']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['3']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['3']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['3']['relative_stop']);
 
         // Assert R4 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['4']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['4']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['4']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['4']['relative_start']);
         // Assert R4 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['4']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['4']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['4']['relative_stop']);
 
         // Assert R5 starts around 10 and not before.
-        $this->assertGreaterThanOrEqual(10, $request_stats['5']['start_diff']);
-        $this->assertLessThanOrEqual(10.5, $request_stats['5']['start_diff']);
+        $this->assertGreaterThanOrEqual(10, $request_stats['5']['relative_start']);
+        $this->assertLessThanOrEqual(10.5, $request_stats['5']['relative_start']);
         // Assert R5 ends around 12.
-        $this->assertGreaterThanOrEqual(11.8, $request_stats['5']['stop_diff']);
-        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['5']['stop_diff']);
+        $this->assertGreaterThanOrEqual(11.8, $request_stats['5']['relative_stop']);
+        $this->assertLessThanOrEqual(12.5 + 1, $request_stats['5']['relative_stop']);
     }
 
     public function testSetRateLimitPerSecondOnePerSecond()
     {
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setRateLimit('1/1s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4338,42 +4366,31 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8002'));
         $multi_curl->addGet(Test::getTestUrl('8003'));
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
 
         // Assert R1 starts around 1 and not before.
-        $this->assertGreaterThanOrEqual(1, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(1.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(1, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(1.5, $request_stats['1']['relative_start']);
 
         // Assert R2 starts around 2 and not before.
-        $this->assertGreaterThanOrEqual(2, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(2.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(2, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(2.5, $request_stats['2']['relative_start']);
     }
 
-    public function testSetRateLimitTenPerThirtySecond()
+    public function testSetRateLimitFivePerThirtySecond()
     {
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('5/30s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4387,34 +4404,23 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=20');
         $multi_curl->addGet(Test::getTestUrl('8002') . '?seconds=10');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R5 starts around 30 and not before.
-        $this->assertGreaterThanOrEqual(30, $request_stats['5']['start_diff']);
-        $this->assertLessThanOrEqual(30.5, $request_stats['5']['start_diff']);
+        $this->assertGreaterThanOrEqual(30, $request_stats['5']['relative_start']);
+        $this->assertLessThanOrEqual(30.5, $request_stats['5']['relative_start']);
     }
 
     public function testSetRateLimitOnePerOneMinute()
     {
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('1/1m');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4425,37 +4431,26 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8001') . '?seconds=70');
         $multi_curl->addGet(Test::getTestUrl('8003') . '?seconds=10');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R1 starts around 60 and not before.
-        $this->assertGreaterThanOrEqual(60, $request_stats['1']['start_diff']);
-        $this->assertLessThanOrEqual(60.5, $request_stats['1']['start_diff']);
+        $this->assertGreaterThanOrEqual(60, $request_stats['1']['relative_start']);
+        $this->assertLessThanOrEqual(60.5, $request_stats['1']['relative_start']);
         // Assert R2 starts around 120 and not before.
-        $this->assertGreaterThanOrEqual(120, $request_stats['2']['start_diff']);
-        $this->assertLessThanOrEqual(120.5, $request_stats['2']['start_diff']);
+        $this->assertGreaterThanOrEqual(120, $request_stats['2']['relative_start']);
+        $this->assertLessThanOrEqual(120.5, $request_stats['2']['relative_start']);
     }
 
     public function testSetRateLimitThreePerOneMinute()
     {
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('3/1m');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4467,34 +4462,23 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8001') . '?seconds=45');
         $multi_curl->addGet(Test::getTestUrl('8003') . '?seconds=10');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R3 starts around 60 and not before.
-        $this->assertGreaterThanOrEqual(60, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(60.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(60, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(60.5, $request_stats['3']['relative_start']);
     }
 
     public function testSetRateLimitThreePerSixtyFiveSeconds()
     {
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('3/65s');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4506,34 +4490,23 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         $multi_curl->addGet(Test::getTestUrl('8003') . '?seconds=5');
         $multi_curl->addGet(Test::getTestUrl('8004') . '?seconds=5');
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // Assert R3 starts around 65 and not before.
-        $this->assertGreaterThanOrEqual(65, $request_stats['3']['start_diff']);
-        $this->assertLessThanOrEqual(65.5, $request_stats['3']['start_diff']);
+        $this->assertGreaterThanOrEqual(65, $request_stats['3']['relative_start']);
+        $this->assertLessThanOrEqual(65.5, $request_stats['3']['relative_start']);
     }
 
     public function testSetRateLimitTenPerTwoMinutes()
     {
-        $request_stats = array(
-            'start' => '',
-            'stop' => '',
-        );
+        $request_stats = [];
 
         $multi_curl = new MultiCurl();
         $multi_curl->setHeader('X-DEBUG-TEST', 'timeout');
         $multi_curl->setRateLimit('10/2m');
         $multi_curl->beforeSend(function ($instance) use (&$request_stats) {
-            $request_stats[$instance->id] = array();
+            $request_stats[$instance->id] = [];
             $request_stats[$instance->id]['start'] = microtime(true);
         });
         $multi_curl->complete(function ($instance) use (&$request_stats) {
@@ -4544,16 +4517,8 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
             $multi_curl->addGet(Test::TEST_URL . '?seconds=1');
         }
 
-        $request_stats['start'] = microtime(true);
         $multi_curl->start();
-        $request_stats['stop'] = microtime(true);
-
-        foreach ($request_stats as $key => &$value) {
-            if (is_int($key)) {
-                $value['start_diff'] = $value['start'] - $request_stats['start'];
-                $value['stop_diff'] = $value['stop'] - $request_stats['start'];
-            }
-        }
+        $request_stats = \Helper\get_request_stats($request_stats, $multi_curl);
 
         // 0-9 starts >= 0.
         // 10-19 starts >= 120.
@@ -4561,35 +4526,35 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
         // 30-39 starts >= 360.
 
         // Assert R0 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['0']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['0']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['0']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['0']['relative_start']);
         // Assert R9 starts around 0 and not before.
-        $this->assertGreaterThanOrEqual(0, $request_stats['9']['start_diff']);
-        $this->assertLessThanOrEqual(0.5, $request_stats['9']['start_diff']);
+        $this->assertGreaterThanOrEqual(0, $request_stats['9']['relative_start']);
+        $this->assertLessThanOrEqual(0.5, $request_stats['9']['relative_start']);
 
         // Assert R10 starts around 120 and not before.
-        $this->assertGreaterThanOrEqual(120, $request_stats['10']['start_diff']);
-        $this->assertLessThanOrEqual(120.5, $request_stats['10']['start_diff']);
+        $this->assertGreaterThanOrEqual(120, $request_stats['10']['relative_start']);
+        $this->assertLessThanOrEqual(120.5, $request_stats['10']['relative_start']);
         // Assert R19 starts around 120 and not before.
-        $this->assertGreaterThanOrEqual(120, $request_stats['19']['start_diff']);
-        $this->assertLessThanOrEqual(120.5, $request_stats['19']['start_diff']);
+        $this->assertGreaterThanOrEqual(120, $request_stats['19']['relative_start']);
+        $this->assertLessThanOrEqual(120.5, $request_stats['19']['relative_start']);
 
         // Assert R20 starts around 240. Allow for some drift.
-        $this->assertGreaterThanOrEqual(239, $request_stats['20']['start_diff']);
-        $this->assertLessThanOrEqual(241, $request_stats['20']['start_diff']);
+        $this->assertGreaterThanOrEqual(239, $request_stats['20']['relative_start']);
+        $this->assertLessThanOrEqual(241, $request_stats['20']['relative_start']);
         // Assert R29 starts around 240. Allow for some drift.
-        $this->assertGreaterThanOrEqual(239, $request_stats['29']['start_diff']);
-        $this->assertLessThanOrEqual(241, $request_stats['29']['start_diff']);
+        $this->assertGreaterThanOrEqual(239, $request_stats['29']['relative_start']);
+        $this->assertLessThanOrEqual(241, $request_stats['29']['relative_start']);
 
         // Assert R30 starts around 360. Allow for some drift.
-        $this->assertGreaterThanOrEqual(359, $request_stats['30']['start_diff']);
-        $this->assertLessThanOrEqual(361, $request_stats['30']['start_diff']);
+        $this->assertGreaterThanOrEqual(359, $request_stats['30']['relative_start']);
+        $this->assertLessThanOrEqual(361, $request_stats['30']['relative_start']);
     }
 
     public function testSetHeadersAssociativeArray()
     {
         $multi_curl = new MultiCurl();
-        $multi_curl->setHeaders(array(
+        $multi_curl->setHeaders([
             ' Key1 ' => ' Value1 ',
             ' Key2 ' => ' Value2',
             ' Key3 ' => 'Value3 ',
@@ -4606,12 +4571,12 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
             'Key14' => ' Value14',
             'Key15' => 'Value15 ',
             'Key16' => 'Value16',
-        ));
+        ]);
         $multi_curl->addGet(Test::TEST_URL);
 
         $curls = \Helper\get_multi_curl_property_value($multi_curl, 'curls');
         foreach ($curls as $curl) {
-            $this->assertEquals(array(
+            $this->assertEquals([
                 'Key1: Value1',
                 'Key2: Value2',
                 'Key3: Value3',
@@ -4628,7 +4593,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
                 'Key14: Value14',
                 'Key15: Value15',
                 'Key16: Value16',
-            ), $curl->getOpt(CURLOPT_HTTPHEADER));
+            ], $curl->getOpt(CURLOPT_HTTPHEADER));
 
             $headers = \Helper\get_curl_property_value($curl, 'headers');
             $this->assertEquals('Value1', $headers['Key1']);
@@ -4653,7 +4618,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
     public function testSetHeadersIndexedArray()
     {
         $multi_curl = new MultiCurl();
-        $multi_curl->setHeaders(array(
+        $multi_curl->setHeaders([
             ' Key1 : Value1 ',
             ' Key2 : Value2',
             ' Key3 :Value3 ',
@@ -4670,12 +4635,12 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
             'Key14: Value14',
             'Key15:Value15 ',
             'Key16:Value16',
-        ));
+        ]);
         $multi_curl->addGet(Test::TEST_URL);
 
         $curls = \Helper\get_multi_curl_property_value($multi_curl, 'curls');
         foreach ($curls as $curl) {
-            $this->assertEquals(array(
+            $this->assertEquals([
                 'Key1: Value1',
                 'Key2: Value2',
                 'Key3: Value3',
@@ -4692,7 +4657,7 @@ class MultiCurlTest extends \PHPUnit\Framework\TestCase
                 'Key14: Value14',
                 'Key15: Value15',
                 'Key16: Value16',
-            ), $curl->getOpt(CURLOPT_HTTPHEADER));
+            ], $curl->getOpt(CURLOPT_HTTPHEADER));
 
             $headers = \Helper\get_curl_property_value($curl, 'headers');
             $this->assertEquals('Value1', $headers['Key1']);
