@@ -6,9 +6,9 @@ use Curl\Curl;
 // curl --request GET "https://httpbin.org/get?key=value"
 
 $curl = new Curl();
-$curl->get('https://httpbin.org/get', array(
+$curl->get('https://httpbin.org/get', [
     'key' => 'value',
-));
+]);
 
 if ($curl->error) {
     echo 'Error: ' . $curl->errorCode . ': ' . $curl->errorMessage . "\n";
