@@ -5,8 +5,8 @@ use Curl\Curl;
 
 $curl = new Curl('https://httpbin.org/get');
 for ($i = 1; $i <= 10; $i++) {
-    $curl->get(array(
+    $curl->get([
         'page' => $i,
-    ));
+    ]);
     // TODO: Do something with result $curl->response.
 }
