@@ -10,14 +10,14 @@ EOF;
 
 $curl = new Curl();
 $curl->setHeader('Content-Type', 'application/json');
-$curl->post('https://api.github.com/gists', array(
+$curl->post('https://api.github.com/gists', [
     'description' => 'PHP-Curl-Class test.',
     'public' => 'true',
-    'files' => array(
-        'Untitled.php' => array(
+    'files' => [
+        'Untitled.php' => [
             'content' => $content,
-        ),
-    ),
-));
+        ],
+    ],
+]);
 
 echo 'Gist created at ' . $curl->response->html_url . "\n";
