@@ -13,18 +13,18 @@ if (isset($_GET['code'])) {
 
     // Exchange the authorization code for an access token.
     $curl = new Curl();
-    $curl->post('https://accounts.google.com/o/oauth2/token', array(
+    $curl->post('https://accounts.google.com/o/oauth2/token', [
         'code' => $code,
         'client_id' => CLIENT_ID,
         'client_secret' => CLIENT_SECRET,
-        'redirect_uri' => implode('', array(
+        'redirect_uri' => implode('', [
             isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? 'https' : 'http',
             '://',
             $_SERVER['SERVER_NAME'],
             $_SERVER['SCRIPT_NAME'],
-        )),
+        ]),
         'grant_type' => 'authorization_code',
-    ));
+    ]);
 
     if ($curl->error) {
         echo $curl->response->error . ': ' . $curl->response->error_description;
@@ -50,18 +50,18 @@ if (isset($_GET['code'])) {
     echo 'Hi ' . $curl->response->displayName . '.';
 } else {
     $curl = new Curl();
-    $curl->get('https://accounts.google.com/o/oauth2/auth', array(
+    $curl->get('https://accounts.google.com/o/oauth2/auth', [
         'scope' => 'https://www.googleapis.com/auth/plus.me',
-        'redirect_uri' => implode('', array(
+        'redirect_uri' => implode('', [
             isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? 'https' : 'http',
             '://',
             $_SERVER['SERVER_NAME'],
             $_SERVER['SCRIPT_NAME'],
-        )),
+        ]),
         'response_type' => 'code',
         'client_id' => CLIENT_ID,
         'approval_prompt' => 'force',
-    ));
+    ]);
 
     $url = $curl->responseHeaders['Location'];
     echo '<a href="' . $url . '">Continue</a>';
