@@ -8,24 +8,24 @@ const INSTAGRAM_CLIENT_SECRET = 'XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX';
 
 session_start();
 
-$redirect_uri = implode('', array(
+$redirect_uri = implode('', [
     isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? 'https' : 'http',
     '://',
     $_SERVER['SERVER_NAME'],
     $_SERVER['SCRIPT_NAME'],
-));
+]);
 
 if (isset($_GET['code'])) {
     $code = $_GET['code'];
 
     $curl = new Curl();
-    $curl->post('https://api.instagram.com/oauth/access_token', array(
+    $curl->post('https://api.instagram.com/oauth/access_token', [
         'client_id' => INSTAGRAM_CLIENT_ID,
         'client_secret' => INSTAGRAM_CLIENT_SECRET,
         'grant_type' => 'authorization_code',
         'redirect_uri' => $redirect_uri,
         'code' => $code,
-    ));
+    ]);
 
     if ($curl->error) {
         echo $curl->response->error_type . ': ' . $curl->response->errorMessage . '<br />';
@@ -38,9 +38,9 @@ if (isset($_GET['code'])) {
 
 if (isset($_SESSION['access_token'])) {
     $curl = new Curl();
-    $curl->get('https://api.instagram.com/v1/media/popular', array(
+    $curl->get('https://api.instagram.com/v1/media/popular', [
         'access_token' => $_SESSION['access_token'],
-    ));
+    ]);
     foreach ($curl->response->data as $media) {
         echo
             '<a href="' . $media->link . '" target="_blank">' .
@@ -48,10 +48,10 @@ if (isset($_SESSION['access_token'])) {
             '</a>';
     }
 } else {
-    header('Location: https://api.instagram.com/oauth/authorize/?' . http_build_query(array(
+    header('Location: https://api.instagram.com/oauth/authorize/?' . http_build_query([
         'client_id' => INSTAGRAM_CLIENT_ID,
         'redirect_uri' => $redirect_uri,
         'response_type' => 'code',
-    )));
+    ]));
     exit;
 }
