@@ -6,11 +6,11 @@ use Curl\Curl;
 const INSTAGRAM_CLIENT_ID = 'XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX';
 
 $curl = new Curl();
-$curl->get('https://api.instagram.com/v1/media/search', array(
+$curl->get('https://api.instagram.com/v1/media/search', [
     'client_id' => INSTAGRAM_CLIENT_ID,
     'lat' => '37.8296',
     'lng' => '-122.4832',
-));
+]);
 
 foreach ($curl->response->data as $media) {
     $image = $media->images->low_resolution;
