@@ -9,9 +9,9 @@ $MAILCHIMP_BASE_URL = 'https://' . $parts['1'] . '.api.mailchimp.com/2.0/';
 
 
 $curl = new Curl();
-$curl->get($MAILCHIMP_BASE_URL . '/lists/list.json', array(
+$curl->get($MAILCHIMP_BASE_URL . '/lists/list.json', [
     'apikey' => MAILCHIMP_API_KEY,
-));
+]);
 
 if ($curl->response->total === 0) {
     echo 'No lists found';
@@ -21,13 +21,13 @@ if ($curl->response->total === 0) {
 $lists = $curl->response->data;
 $list = $lists['0'];
 
-$curl->post($MAILCHIMP_BASE_URL . '/lists/subscribe.json', array(
+$curl->post($MAILCHIMP_BASE_URL . '/lists/subscribe.json', [
     'apikey' => MAILCHIMP_API_KEY,
     'id' => $list->id,
-    'email' => array(
+    'email' => [
         'email' => 'user@example.com',
-    ),
-));
+    ],
+]);
 
 if ($curl->error) {
     echo $curl->response->name . ': ' . $curl->response->error . "\n";
