@@ -5,14 +5,14 @@ use Curl\MultiCurl;
 
 $multi_curl = new MultiCurl();
 
-$multi_curl->addGet('https://www.google.com/search', array(
+$multi_curl->addGet('https://www.google.com/search', [
     'q' => 'hello world',
-));
-$multi_curl->addGet('https://duckduckgo.com/', array(
+]);
+$multi_curl->addGet('https://duckduckgo.com/', [
     'q' => 'hello world',
-));
-$multi_curl->addGet('https://www.bing.com/search', array(
+]);
+$multi_curl->addGet('https://www.bing.com/search', [
     'q' => 'hello world',
-));
+]);
 
 $multi_curl->start();
