@@ -18,14 +18,14 @@ $multi_curl->complete(function ($instance) {
     echo 'call to "' . $instance->url . '" completed.' . "\n";
 });
 
-$multi_curl->addGet('https://www.google.com/search', array(
+$multi_curl->addGet('https://www.google.com/search', [
     'q' => 'hello world',
-));
-$multi_curl->addGet('https://duckduckgo.com/', array(
+]);
+$multi_curl->addGet('https://duckduckgo.com/', [
     'q' => 'hello world',
-));
-$multi_curl->addGet('https://www.bing.com/search', array(
+]);
+$multi_curl->addGet('https://www.bing.com/search', [
     'q' => 'hello world',
-));
+]);
 
 $multi_curl->start();
