@@ -5,15 +5,15 @@ use Curl\MultiCurl;
 
 $multi_curl = new MultiCurl();
 
-$multi_curl->addPost('https://httpbin.org/post', array(
+$multi_curl->addPost('https://httpbin.org/post', [
     'to' => 'alice',
     'subject' => 'hi',
     'body' => 'hi Alice',
-));
-$multi_curl->addPost('https://httpbin.org/post', array(
+]);
+$multi_curl->addPost('https://httpbin.org/post', [
     'to' => 'bob',
     'subject' => 'hi',
     'body' => 'hi Bob',
-));
+]);
 
 $multi_curl->start();
