@@ -10,16 +10,16 @@ $multi_curl = new MultiCurl();
 // then try uncommenting the following line:
 // $multi_curl->setHeader('Content-Type', 'multipart/form-data');
 
-$multi_curl->addPost('https://httpbin.org/post', array(
+$multi_curl->addPost('https://httpbin.org/post', [
     'image' => new CURLFile('the-lorax.jpg'),
-));
+]);
 
-$multi_curl->addPost('https://httpbin.org/post', array(
+$multi_curl->addPost('https://httpbin.org/post', [
     'image' => new CURLFile('swomee-swans.jpg'),
-));
+]);
 
-$multi_curl->addPost('https://httpbin.org/post', array(
+$multi_curl->addPost('https://httpbin.org/post', [
     'image' => new CURLFile('truffula-trees.jpg'),
-));
+]);
 
 $multi_curl->start();
