@@ -6,9 +6,9 @@ use Curl\Curl;
 // curl --get --request OPTIONS "http://127.0.0.1:8000/" --data "foo=bar"
 
 $curl = new Curl();
-$curl->options('http://127.0.0.1:8000/', array(
+$curl->options('http://127.0.0.1:8000/', [
     'foo' => 'bar',
-));
+]);
 
 if ($curl->error) {
     echo 'Error: ' . $curl->errorCode . ': ' . $curl->errorMessage . "\n";
