@@ -9,11 +9,11 @@ use Curl\Curl;
 //     -d '{"id":"1","content":"Hello world!","date":"2015-06-30 19:42:21"}' \
 //     "https://httpbin.org/post"
 
-$data = json_encode(array(
+$data = json_encode([
     'id' => '1',
     'content' => 'Hello world!',
     'date' => date('Y-m-d H:i:s'),
-), JSON_UNESCAPED_UNICODE);
+], JSON_UNESCAPED_UNICODE);
 
 $curl = new Curl();
 $curl->setHeader('Content-Type', 'application/json');
