@@ -6,11 +6,11 @@ use Curl\Curl;
 // curl --request PUT "https://httpbin.org/put" --data "id=1&first_name=Zach&last_name=Borboa"
 
 $curl = new Curl();
-$curl->put('https://httpbin.org/put', array(
+$curl->put('https://httpbin.org/put', [
     'id' => '1',
     'first_name' => 'Zach',
     'last_name' => 'Borboa',
-));
+]);
 
 if ($curl->error) {
     echo 'Error: ' . $curl->errorCode . ': ' . $curl->errorMessage . "\n";
