@@ -11,8 +11,8 @@ $curl = new Curl('https://www.example.com/search');
 
 for ($i = 1; $i <= $pages; $i++) {
     // https://www.example.com/search?q=coffee&page=N
-    $curl->get(array(
+    $curl->get([
         'q' => $q,
         'page' => $i,
-    ));
+    ]);
 }
