@@ -20,21 +20,9 @@ class Encoder
     public static function encodeJson()
     {
         $args = func_get_args();
-
-        // Call json_encode() without the $depth parameter in PHP
-        // versions less than 5.5.0 as the $depth parameter was added in
-        // PHP version 5.5.0.
-        if (version_compare(PHP_VERSION, '5.5.0', '<')) {
-            $args = array_slice($args, 0, 2);
-        }
-
         $value = call_user_func_array('json_encode', $args);
         if (json_last_error() !== JSON_ERROR_NONE) {
-            if (function_exists('json_last_error_msg')) {
-                $error_message = 'json_encode error: ' . json_last_error_msg();
-            } else {
-                $error_message = 'json_encode error';
-            }
+            $error_message = 'json_encode error: ' . json_last_error_msg();
             throw new \ErrorException($error_message);
         }
         return $value;
